Create doctor dashboard and routing
