+++ Clinic/Clinic/Views/Shared/_Layout.cshtml
@@ -82,6 +82,13 @@
                         {
                             <li class="nav-item">
                                 <a class="nav-link" asp-area="Doctor"
+                                   asp-controller="Doctor"
+                                   asp-action="Index">
+                                    Dashboard
+                                </a>
+                            </li>
+                            <li class="nav-item">
+                                <a class="nav-link" asp-area="Doctor"
                                    asp-controller="DoctorAppointments"
                                    asp-action="Index">
                                     Moje wizyty
